refactor: Cria hook para listar especialidades
